refactor: 엔티티들에 BaseEntity 상속 적용 및 User 도메인 공통 패키징 구조 개선
