+++ src/main/java/com/example/outsourcing_project/domain/auth/domain/refresh/RefreshToken.java
@@ -1,6 +1,6 @@
 package com.example.outsourcing_project.domain.auth.domain.refresh;
 
-import com.example.outsourcing_project.domain.user.domain.User;
+import com.example.outsourcing_project.domain.user.domain.model.User;
 import jakarta.persistence.*;
 import lombok.*;
