+++ src/main/java/com/example/outsourcing_project/domain/auth/domain/refresh/RefreshTokenService.java
@@ -1,6 +1,6 @@
 package com.example.outsourcing_project.domain.auth.domain.refresh;
 
-import com.example.outsourcing_project.domain.user.domain.User;
+import com.example.outsourcing_project.domain.user.domain.model.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import java.time.Instant;
