+++ src/main/java/com/example/outsourcing_project/domain/comments/controller/CommentCreateResponseDto.java
@@ -2,7 +2,7 @@ package com.example.outsourcing_project.domain.comments.controller;
 
 import com.example.outsourcing_project.domain.comments.model.entity.Comments;
 import com.example.outsourcing_project.domain.task.domain.model.Task;
-import com.example.outsourcing_project.domain.user.domain.User;
+import com.example.outsourcing_project.domain.user.domain.model.User;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
