+++ src/main/java/com/example/outsourcing_project/domain/comments/model/entity/Comments.java
@@ -1,7 +1,7 @@
 package com.example.outsourcing_project.domain.comments.model.entity;
 
 import com.example.outsourcing_project.domain.task.domain.model.Task;
-import com.example.outsourcing_project.domain.user.domain.User;
+import com.example.outsourcing_project.domain.user.domain.model.User;
 import com.example.outsourcing_project.global.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
