+++ src/main/java/com/example/outsourcing_project/domain/log/domain/model/Log.java
@@ -1,6 +1,5 @@
 package com.example.outsourcing_project.domain.log.domain.model;
 
-import com.example.outsourcing_project.domain.user.domain.User;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
