+++ src/main/java/com/example/outsourcing_project/domain/log/service/LogSaveDto.java
@@ -1,7 +1,6 @@
 package com.example.outsourcing_project.domain.log.service;
 
 import com.example.outsourcing_project.domain.log.domain.model.Log;
-import com.example.outsourcing_project.domain.user.domain.User;
 import lombok.Getter;
 import lombok.Setter;
