+++ src/main/java/com/example/outsourcing_project/domain/user/domain/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.example.outsourcing_project.domain.user.domain;
+package com.example.outsourcing_project.domain.user.domain.repository;
 
+import com.example.outsourcing_project.domain.user.domain.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
